--- get_feature_properties/beaches.json
@@ -1,48 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -80.12603759765625,
-              25.760628957641366
-            ],
-            [
-              -80.11608123779297,
-              25.82863401406832
-            ],
-            [
-              -80.11505126953125,
-              25.932112145619172
-            ],
-            [
-              -80.15865325927734,
-              25.930877120921785
-            ],
-            [
-              -80.18646240234375,
-              25.843157307531634
-            ],
-            [
-              -80.19401550292969,
-              25.780416269800465
-            ],
-            [
-              -80.16345977783203,
-              25.762793355586627
-            ],
-            [
-              -80.12603759765625,
-              25.760628957641366
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
